pyproject.toml
[tool.poetry]
name = "mbay-dict"
version = "0.1.0"
description = ""
authors = ["Ndil-Adjim  Sou Ngadoy <nasoungadoy@gmail.com>"]
readme = "README.md"
packages = [{include = "mbay_dict"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.259"
openai = "^0.27.8"
pydantic = "==1.*"
httpx = "^0.24.1"
tenacity = "^8.2.2"
python-dotenv = "^1.0.0"
more-itertools = "^10.1.0"
jsonlines = "^3.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
jupyterlab = "^4.0.4"
ruff = "^0.0.283"
colorama = "^0.4.6"
rich = "^13.5.2"
ipython = "^8.14.0"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
